soft delete instead of hard delete

+++ hnproj/topstoriesviews.py
@@ -17,6 +17,8 @@ def home(request):
   topStories = HNTopStory.objects.all()
   storyJSONs = []
   for story in topStories:
+    if story.markDeleted is True:
+      continue;
     jsonStory = json.loads(story.story);
     if not jsonStory.get('score'):
       story.delete();
